CLI apps concurrency fix
