+++ build.gradle.kts
@@ -36,7 +36,7 @@ repositories {
 dependencies {
     implementation(gradleApi())
 
-    implementation("com.cognifide.gradle:common-plugin:0.1.47")
+    implementation("com.cognifide.gradle:common-plugin:0.1.48")
 
     implementation("org.jsoup:jsoup:1.12.1")
     implementation("org.buildobjects:jproc:2.3.0")
