+++ gradle.properties
@@ -1,3 +1,3 @@
-version=13.0.0-beta
+version=13.0.0-beta2
 release.useAutomaticVersion=true
 org.gradle.jvmargs=-Xmx1024m
