+++ src/main/kotlin/com/cognifide/gradle/aem/common/instance/LocalInstanceManager.kt
@@ -10,6 +10,7 @@ import com.cognifide.gradle.aem.javaVersions
 import com.cognifide.gradle.common.pluginProject
 import com.cognifide.gradle.common.utils.onEachApply
 import com.cognifide.gradle.common.utils.using
+import kotlinx.coroutines.withTimeout
 import org.buildobjects.process.ProcBuilder
 import org.gradle.api.JavaVersion
 import java.io.File
@@ -393,7 +394,7 @@ class LocalInstanceManager(internal val aem: AemExtension) : Serializable {
             common.parallel.with(upInstances) {
                 increment("Opening instance '$name'") {
                     try {
-                        executeOpenScript()
+                        aem.webBrowser.open(httpOpenUrl) { withTimeoutMillis(openTimeout.get()) }
                         openedInstances += this@with
                     } catch (e: LocalInstanceException) {
                         logger.debug("Instance '$name' open error", e)
